Make istioctl kube client common (#24650)
There is a lot of redundant code between the istioctl kube client and the testing kube Accessor. This PR attempts to define a unified client that can be shared across the codebase. For now, it is (mostly) a repackaging of the istioctl kubernetes.Client with a few minor tweaks from Accessor. In particular, I've combined the two interfaces into a single Client interface.

Follow on work will remove most (if not all) of the test kube Accessor and replace it with this new client.