Fix breakage in istioctl manifest apply (#25451)
This was introduced by #25384. The issue was that the parameter order was swapped in order to be consistent with GenerateConfig (and a few other places). There was apparently no test coverage for this beyond the integration test framework's usage when installing istio. Unfortunately, we don't want to use the commands directly there because the use of flags interferes with the ability to install multiple control planes in parallel.

Fixing the issue by swapping to a structure for the parameter to avoid similar breakage in the future.

Fixes #25447