Fix webhook tests for rewrite probe (#28975)
In https://github.com/istio/istio/pull/28800, we changed to reading the
probe setting indirectly from the injection template to reading it
directly from the source (values). This made these tests, which have a
custom template and don't set the values, not test the right thing.

This updates the tests to set the values correctly. Additionally, I
removed the obsolete field in injector template.

An alternative to this PR would be to continue to read the setting from
the template. I do not think its a better option, but don't feel
strongly either way; it should not impact users either way