injection: remove attempted legacy injection template support (#29652)
* injection: remove attempted legacy injection template support

I originally added this in an attempt to support reading 1.8 template by
1.9 istiod. This is *not* something we have backwards compatibility
guarantees, and its something we have broken almost every release.

I later realized its not feasible to support this, and since its not
even very useful (it mostly helps when someone accidentally uses the
`latest` image but release charts - which is good to fail fast). So this
PR cleans it up.

* update test