Remove local XDS generator (#29795)
* Remove local XDS generator

This code was added as an experiment for local proxy generation. It is
not currently used, but is adding a fair amount of complexity to support
the tests and authentication logic. It also requires linking in the full
XDS libraries to work, which have been disabled in the agent. I think it would make sense to
continue experimentation in a separate branch if desired, and un-burden
master while experiments go on.

* Cleanup more test code

* Fix deadlock

* remove howardjohn

* lint

* revert wait change